+++ tests/estimation/test_estimation_factories.py
@@ -86,9 +86,25 @@ def testSequentialFilterFactoryError(dynamics):
     config["name"] = "ukf"
     # Create config object
     estimation_config = SequentialFilterConfig(**config)
-    estimation_config.name = "invlaid_name"
+    estimation_config.name = "invalid_name"
     # Call factory function
     error_msg = f"Invalid filter type: {estimation_config.name}"
+    with pytest.raises(ValueError, match=error_msg):
+        _ = sequentialFilterFactory(
+            estimation_config, TGT_ID, INITIAL_TIME, EST_X, EST_P, dynamics, Q_MATRIX
+        )
+
+
+def testSequentialFilterFactoryDupManeuverHandlingError(dynamics):
+    """Tests catching errors for setting MMAE and IOD for a single filter."""
+    config = deepcopy(ESTIMATION_CONFIG["sequential_filter"])
+    config["name"] = "ukf"
+    config["adaptive_estimation"] = True
+    config["initial_orbit_determination"] = True
+    # Create config object
+    estimation_config = SequentialFilterConfig(**config)
+    # Call factory function
+    error_msg = "IOD & MMAE cannot be used at the same time"
     with pytest.raises(ValueError, match=error_msg):
         _ = sequentialFilterFactory(
             estimation_config, TGT_ID, INITIAL_TIME, EST_X, EST_P, dynamics, Q_MATRIX